Respond to final review comment, add series description setting for export.
